// src/app/enums/date-range.ts
export enum DateRangeType {
    Today = 'Today',
    Yesterday = 'Yesterday',
    Last5Days = 'Last 5 Days',
    Last7Days = 'Last 7 Days',
    Last14Days = 'Last 14 Days',
    Last30Days = 'Last 30 Days',
    Last60Days = 'Last 60 Days',
    Last90Days = 'Last 90 Days',
    LastWeek = 'Last Week',
    Last4Weeks = 'Last 4 Weeks',
    Last13Weeks = 'Last 13 Weeks',
    Last52Weeks = 'Last 52 Weeks',
    ThisMonth = 'This Month',
    LastMonth = 'Last Month',
    Last3Months = 'Last 3 Months',
    Last6Months = 'Last 6 Months',
    Last12Months = 'Last 12 Months',
    YTD = 'Year to date',
    FYTD = 'Fiscal Year to date',
    customRange = 'Custom Range',
    LatestWeek = 'Latest Week',
    LastestMonth = 'Latest Month',
    previousWeek = 'Previous Week',
    previousMonth = 'Previous Month',
  
  }

  export enum IntervalType {
    hourly = 'HOURLY',
    daily = 'DAILY',
    weekly = 'WEEKLY',
    monthly = 'MONTHLY',
    quaterly = 'QUARTERLY',
    yearly = 'YEARLY'
  }
  
  export enum Interval {
    day = 'Day',
    week = 'Week',
    month = 'Month',
  }